Fix explorer minimize hotkey not working correctly in VDI - it's because the no-machine-filter row was listed first.
Since the first row wins when getting rows by column, it was winning.
It's not filtered out because the default machine filters always allow blanks through.

Could fix it by moving more specific (machine-filtered) line above the other, but it doesn't make sense to enforce an order like that. Instead, we're now using the machine column as a tiebreaker so that the proper one wins even if it comes after (same as programs TL handling does).