Temporary commit: saving off how far I got with generating a function header by parsing the parameters out of a function signature.
Plan was to get the cursor to the start of the line with AHK, then have the snippet include a newline (since we can't look at the next line from a snippet).

Regex transformation currently extracts each of the parameters to the function, with group 2 telling us whether it's optional (if it exists).